refactor: restructure backend directory for modular architecture
