+++ backend/.env.example
@@ -1,0 +1,11 @@
+# Server Configuration
+PORT=5000
+
+# JWT Secret Key (Change this to a strong, random string in production)
+JWT_SECRET=your-super-secret-jwt-key-change-this-in-production
+
+# Environment
+NODE_ENV=development
+
+# CORS Origins (comma-separated list of allowed origins)
+CORS_ORIGINS=http://localhost:3000,http://localhost:3001
