+++ backend/config/environment.js
@@ -1,0 +1,8 @@
+require('dotenv').config();
+
+module.exports = {
+  PORT: process.env.PORT || 5000,
+  JWT_SECRET: process.env.JWT_SECRET || 'your-secret-key',
+  NODE_ENV: process.env.NODE_ENV || 'development',
+  CORS_ORIGINS: process.env.CORS_ORIGINS?.split(',') || ['http://localhost:3000']
+};
