+++ backend/middlewares/errorHandler.js
@@ -1,0 +1,10 @@
+const errorHandler = (error, req, res, next) => {
+  console.error('Unhandled error:', error);
+  res.status(500).json({ message: 'Internal server error' });
+};
+
+const notFoundHandler = (req, res) => {
+  res.status(404).json({ message: 'Route not found' });
+};
+
+module.exports = { errorHandler, notFoundHandler };
