+++ backend/routes/attendance.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const { authenticateToken } = require('../middlewares/auth');
+const { getAttendance, updateAttendance } = require('../controllers/attendanceController');
+
+const router = express.Router();
+
+router.get('/', authenticateToken, getAttendance);
+router.post('/', authenticateToken, updateAttendance);
+
+module.exports = router;
